Possible to send friendrequests
Okay this is really strange

I added the host for the kwick mobile version to the api
because some functions work there. I'm pretty sure the real MAPI
has a way to send friend requests as well, but the docs are wrong
and I have no idea what the correct url for this in MAPI 1.0
is.
This adds a second session to the library as well, one with the
session from mapi and one with the mobile version session.